Pooly: v4: Link Workers to Server (!!!)
This is so important I have put it in its own commit.
After trying to test the blocking queue feature I have seen PoolyServer
does not notice when a worker :EXIT, even from a :normal reason.
That is because the code has a bug. It 'unlink' the Server from the worker
when it 'dissmiss' the worker after it ends the work but Server has never
been linked to the worker to begin with ...