hangman: Refactor common types
Some types were used both by the API and the implementation.
That leads to a circular dependency that, although properly
handled by Elixir, is a design smell.
For that reason we have created a new types file with all
the shared types (just two for now) and reference it in
both API and implementation code.