hangman: html_client: Use PRG
POST-REDIRECT-GET
Whenever you hit reload in a browser it resend
the last request.

If the last request is a POST/PUT the request to change
the server status will be sent again and that could cause problems.

To avoid this problem, servers *DO NOT RENDER PAGES ON PUT/POST*
instead they redirect the browser to another page where
the result of the change is shown. This way the browser
send a GET as the last request and any RELOAD will only
ask for the same status not triggering any duplicated change.

Signed-off-by: Jose Luis Ojosnegros Manchón <jl.ojosnegros.manchon@gmail.com>